+++ OpenOversight/app/models.py
@@ -1,11 +1,9 @@
-import base64
 import re
 import time
 from datetime import date
-from email.mime.text import MIMEText
 
 from authlib.jose import JoseError, JsonWebToken
-from flask import current_app, render_template
+from flask import current_app
 from flask_login import UserMixin
 from flask_sqlalchemy import SQLAlchemy
 from flask_sqlalchemy.model import DefaultMeta
@@ -622,65 +620,3 @@ class User(UserMixin, BaseModel):
 @login_manager.user_loader
 def load_user(user_id):
     return User.query.get(int(user_id))
-
-
-class Email:
-    """Base class for all emails."""
-
-    def __init__(self, body: str, subject: str, receiver: str):
-        self.body = body
-        self.receiver = receiver
-        self.subject = subject
-
-    def create_message(self):
-        message = MIMEText(self.body, "html")
-        message["to"] = self.receiver
-        message["from"] = current_app.config["OO_SERVICE_EMAIL"]
-        message["subject"] = self.subject
-        return {"raw": base64.urlsafe_b64encode(message.as_bytes()).decode()}
-
-
-class AdministratorApprovalEmail(Email):
-    def __init__(self, receiver: str, user, admin):
-        subject = (
-            f"{current_app.config.get('OO_MAIL_SUBJECT_PREFIX')} New User Registered"
-        )
-        body = render_template("auth/email/new_registration.html", user, admin)
-        super().__init__(body, subject, receiver)
-
-
-class ChangeEmailAddressEmail(Email):
-    def __init__(self, receiver: str, user, token: str):
-        subject = (
-            f"{current_app.config.get('OO_MAIL_SUBJECT_PREFIX')} Confirm Your Email "
-            f"Address"
-        )
-        body = render_template("auth/email/change_email.html", user, token)
-        super().__init__(body, subject, receiver)
-
-
-class ConfirmAccountEmail(Email):
-    def __init__(self, receiver: str, user, token: str):
-        subject = (
-            f"{current_app.config.get('OO_MAIL_SUBJECT_PREFIX')} Confirm Your Account"
-        )
-        body = render_template("auth/email/confirm.html", user, token)
-        super().__init__(body, subject, receiver)
-
-
-class ConfirmedUserEmail(Email):
-    def __init__(self, receiver: str, user, admin):
-        subject = (
-            f"{current_app.config.get('OO_MAIL_SUBJECT_PREFIX')} New User Confirmed"
-        )
-        body = render_template("auth/email/new_confirmation.html", user, admin)
-        super().__init__(body, subject, receiver)
-
-
-class ResetPasswordEmail(Email):
-    def __init__(self, receiver: str, user, token: str):
-        subject = (
-            f"{current_app.config.get('OO_MAIL_SUBJECT_PREFIX')} Reset Your Password"
-        )
-        body = render_template("auth/email/reset_password.html", user, token)
-        super().__init__(body, subject, receiver)
